[refactor] DDD 구조에 맞게 수정
