+++ hub-service/src/main/java/com/vroomvroom/hub/application/HubServiceImpl.java
@@ -2,7 +2,7 @@ package com.vroomvroom.hub.application;
 
 import com.vroomvroom.common.api.PageResponse;
 import com.vroomvroom.hub.application.command.CreateHubCommand;
-import com.vroomvroom.hub.application.dto.CreateHubRes;
+import com.vroomvroom.hub.presentation.dto.response.CreateHubRes;
 import com.vroomvroom.hub.application.dto.HubDetailRes;
 import com.vroomvroom.hub.application.dto.HubListRes;
 import com.vroomvroom.hub.domain.entity.Hub;
